+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+v1.5.1
+---------
+2024-02-28:
+    - overhaul documentation
+    - github actions/checkout@v4
+    - github actions/setup-python@v5
+
 v1.5.0
 ---------
 2024-02-27: